Move shared maintenance window code to plugin
Use this plugin in PostgresResource and Vm.
