+++ model/postgres/postgres_resource.rb
@@ -28,6 +28,7 @@ class PostgresResource < Sequel::Model
     :storage_auto_scale_action_performed_80, :storage_auto_scale_action_performed_85, :storage_auto_scale_action_performed_90,
     :storage_auto_scale_canceled, :storage_auto_scale_not_cancellable, :skip_strict_memory_overcommit,
     :bypass_maintenance_window
+  plugin MaintenanceWindow
   include ObjectTag::Cleanup
 
   ServerExclusionFilters = Struct.new(:exclude_host_ids, :exclude_data_centers, :exclude_availability_zones, :availability_zone)
@@ -231,18 +232,6 @@ class PostgresResource < Sequel::Model
     effective_timeline.latest_backup_size_in_gib > target_storage_size_gib
   end
 
-  def in_maintenance_window?
-    return true if maintenance_window_start_at.nil?
-    now = Time.now.utc
-    offset = (now.hour - maintenance_window_start_at) % 24
-    return false unless offset < MAINTENANCE_DURATION_IN_HOURS
-    return true if maintenance_window_days_bitmask.zero?
-
-    # Check the day the window opened at, in case window crosses the midnight
-    bit = ((now - offset * 3600).wday + 6) % 7
-    maintenance_window_days_bitmask[bit] == 1
-  end
-
   def pending_platform_maintenance?
     !Semaphore.where(strand_id: servers_dataset.select(:id), name: "recycle").empty?
   end
@@ -330,22 +319,6 @@ class PostgresResource < Sequel::Model
       root_cert_key = OpenSSL::PKey::EC.new(key_pem_2)
     end
     [root_cert, root_cert_key]
-  end
-
-  def validate
-    super
-    validates_includes(0..23, :maintenance_window_start_at, allow_nil: true, message: "must be between 0 and 23")
-    validates_includes(0..127, :maintenance_window_days_bitmask, allow_nil: true, message: "must be between 0 and 127")
-  end
-
-  # An empty list means the window applies every day (the bitmask is 0).
-  def maintenance_window_day_names
-    DAYS_OF_WEEK.each_with_index.filter_map { |d, i| d if maintenance_window_days_bitmask[i] == 1 }
-  end
-
-  # [day, checked] pairs for the maintenance window days.
-  def maintenance_window_day_options_state
-    DAYS_OF_WEEK.each_with_index.map { |day, i| [day, maintenance_window_days_bitmask[i] == 1] }
   end
 
   def read_replica?
@@ -736,28 +709,6 @@ class PostgresResource < Sequel::Model
     # back to the standard list.
     versions = Option::POSTGRES_VERSION_OPTIONS[flavor] || Option::POSTGRES_VERSION_OPTIONS[Flavor::STANDARD]
     versions.include?(DEFAULT_VERSION) ? DEFAULT_VERSION : versions.max_by(&:to_i)
-  end
-
-  MAINTENANCE_DURATION_IN_HOURS = 2
-  DAYS_OF_WEEK = %w[mon tue wed thu fri sat sun].freeze
-  DAYS_OF_WEEK_BIT = DAYS_OF_WEEK.each_with_index.to_h { |d, i| [d, 1 << i] }.freeze
-
-  def self.maintenance_hour_options
-    Array.new(24) do
-      [it, "#{"%02d" % it}:00 - #{"%02d" % ((it + MAINTENANCE_DURATION_IN_HOURS) % 24)}:00 (UTC)"]
-    end
-  end
-
-  def self.maintenance_window_days_mask(names)
-    return 0 if names.nil? || names.empty?
-
-    # Days of week are stored as a bitmask and presented to the user
-    # as list of short strings stored in DAYS_OF_WEEK
-    names.reduce(0) do |mask, name|
-      mask | DAYS_OF_WEEK_BIT.fetch(name.to_s.downcase) do
-        fail Validation::ValidationFailed.new({maintenance_window_days: "\"#{name}\" is not a valid day. Valid days: #{DAYS_OF_WEEK.join(", ")}"})
-      end
-    end
   end
 
   UPGRADE_IMAGE_MIN_VERSIONS = {
