+++ model/vm.rb
@@ -50,6 +50,7 @@ class Vm < Sequel::Model
   plugin ProviderDispatcher, __FILE__
   plugin SemaphoreMethods, :destroy, :start_after_host_reboot, :prevent_destroy, :update_firewall_rules,
     :checkup, :update_spdk_dependency, :waiting_for_capacity, :lb_expiry_started, :restart, :start, :stop, :migrate_to_separate_progs, :admin_stop, :stopping, :prepare_to_move
+  plugin MaintenanceWindow
   include HealthMonitorMethods
 
   include ObjectTag::Cleanup
@@ -218,44 +219,6 @@ class Vm < Sequel::Model
       retry
     end
     raise
-  end
-
-  def validate
-    super
-    validates_includes(0..23, :maintenance_window_start_at, allow_nil: true, message: "must be between 0 and 23")
-    validates_includes(0..127, :maintenance_window_days_bitmask, allow_nil: true, message: "must be between 0 and 127")
-  end
-
-  MAINTENANCE_DURATION_IN_HOURS = 2
-  DAYS_OF_WEEK = %w[mon tue wed thu fri sat sun].freeze
-  DAYS_OF_WEEK_BIT = DAYS_OF_WEEK.each_with_index.to_h { |d, i| [d, 1 << i] }.freeze
-
-  def self.maintenance_hour_options
-    Array.new(24) do
-      [it, "#{"%02d" % it}:00 - #{"%02d" % ((it + MAINTENANCE_DURATION_IN_HOURS) % 24)}:00 (UTC)"]
-    end
-  end
-
-  def self.maintenance_window_days_mask(names)
-    return 0 if names.nil? || names.empty?
-
-    # Days of week are stored as a bitmask and presented to the user
-    # as list of short strings stored in DAYS_OF_WEEK
-    names.reduce(0) do |mask, name|
-      mask | DAYS_OF_WEEK_BIT.fetch(name.to_s.downcase) do
-        fail Validation::ValidationFailed.new({maintenance_window_days: "\"#{name}\" is not a valid day. Valid days: #{DAYS_OF_WEEK.join(", ")}"})
-      end
-    end
-  end
-
-  # An empty list means the window applies every day (the bitmask is 0).
-  def maintenance_window_day_names
-    DAYS_OF_WEEK.each_with_index.filter_map { |d, i| d if maintenance_window_days_bitmask[i] == 1 }
-  end
-
-  # [day, checked] pairs for the maintenance window days.
-  def maintenance_window_day_options_state
-    DAYS_OF_WEEK.each_with_index.map { |day, i| [day, maintenance_window_days_bitmask[i] == 1] }
   end
 
   include Validation::PublicKeyValidation
